% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN peach.R
\name{h2o_peach}
\alias{h2o_peach}
\title{DNN peach}
\usage{
h2o_peach(x, y, train.data, test.data, ntrees_opts = c(10000),
  max_depth_opts = seq(1, 20), min_rows_opts = c(1, 5, 10, 20, 50, 100),
  learn_rate_opts = seq(0.001, 0.01, 0.001), sample_rate_opts = seq(0.3, 1,
  0.05), col_sample_rate_opts = seq(0.3, 1, 0.05),
  col_sample_rate_per_tree_opts = seq(0.3, 1, 0.05))
}
\description{
DNN peach
}
