% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peach.r
\name{PEACH}
\alias{PEACH}
\alias{peach_main}
\title{Detection of cell separation–induced gene expression through a penalized deconvolution approach.}
\usage{
peach_main(ref, bulk, iter = 10^4, parC = 10^-4)
}
\arguments{
\item{iter}{The number of iterations. Default is 10000.}

\item{parC}{A parameter of the spike-and-slab distribution. Default is 10^-4.}

\item{res.set}{A G by N expression matrix of reference, where G is the gene number and N is the total number of reference.}

\item{mix.set}{A G by M expression matrix of bulk samples, where G is the gene number and M is the total number of bulk samples.}
}
\value{
A list of estimated cell proportions (w),
the probability of informative genes (p),
the fraction of informative genes per cell type (pi), and
the cell-specific profiles (mean).
}
\description{
PEACH adopts a deconvolution method to detect cell separation–induced genes.
}
\author{
An-Shun Tai \email{anshuntai@nctu.edu.tw}
}
