
\name{TCGA.input}
\docType{data}
\alias{TCGA.input}
\title{A sample of 16 cancers gene differential expression analysis t statistic score data}
\description{
  This data set gives the TCGA 16 cancer types tumor versus normal t-test scores of each gene.
}
\usage{TCGA.input
}
\format{A matrix containing the t statistics of 20501 genes/rows and 16 cancer types/columns}
\source{TCGA}
\keyword{datasets}