% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN peach.R
\name{h2o_peach}
\alias{h2o_peach}
\title{DNN peach}
\usage{
h2o_peach(x, y, ntrees_opts = c(10000), max_depth_opts = seq(1, 20),
  min_rows_opts = c(1, 5, 10, 20, 50, 100), learn_rate_opts = seq(0.001,
  0.01, 0.001), sample_rate_opts = seq(0.3, 1, 0.05),
  col_sample_rate_opts = seq(0.3, 1, 0.05),
  col_sample_rate_per_tree_opts = seq(0.3, 1, 0.05))
}
\description{
DNN peach
}
\examples{
set.seed(960806)
sample1<- sample(1:nrow(t.data1),round(nrow(t.data1)*0.8,0))
train.data <- as.h2o(t.data1[sample1,])
test.data <- as.h2o(t.data1[-sample1,])
response1 <- "mean_price"
response2 <- "trade"
x=c("mean_T","min_T","max_T","prec","max_ws","mean_ws","rh","light","mon","weekday","trade",'')
y = response2
h2o_peach(x,y)

gbm_sorted_grid <- h2o.getGrid(grid_id = "mygrid", sort_by = "mse")
print(gbm_sorted_grid)
best_model <- h2o.getModel(gbm_sorted_grid@model_ids[[1]])
summary(best_model)

dnn.pred <- as.numeric(h2o.predict(best_model, test.data))

x<-as.data.frame(test.data$mean_price)
y<-as.data.frame(dnn.pred$predict)

cor(x$mean_price,y$predict)^2
abline(0,1,col="red")
}
