% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNN peach.R
\name{h2o_peach}
\alias{h2o_peach}
\title{DNN peach}
\usage{
h2o_peach(x, y, train.data, test.data, ntrees_opts = c(1000),
  max_depth_opts = seq(1, 10), min_rows_opts = c(1, 5, 10, 20, 50, 100),
  learn_rate_opts = seq(0.001, 0.01, 0.001), sample_rate_opts = seq(0.4, 1,
  0.05), col_sample_rate_opts = seq(0.4, 1, 0.05),
  col_sample_rate_per_tree_opts = seq(0.4, 1, 0.05))
}
\value{
list
}
\description{
DNN peach
}
\examples{
 set.seed(960806)
setwd('../')
a<-paste0(getwd(),'/Desktop')
setwd(a)
library(h2o)
h2o.init()
sample1<- sample(1:nrow(t.data1),round(nrow(t.data1)*0.8,0))
train.data <- as.h2o(t.data1[sample1,])
test.data <- as.h2o(t.data1[-sample1,])
response1 <- "mean_price"
response2 <- "trade"
x=c("mean_T","min_T","max_T","prec","max_ws","mean_ws","rh","light","mon","weekday","trade",'prec1')
y = response2
h2o_peach(x,y,train.data,test.data)
xx<-as.data.frame(test.data$mean_price)
yy<-as.data.frame(dnn.pred$predict)
plot(xx[,1],yy$predict,xlab=names(xx),ylab=names(yy))
abline(0,1,col=2)
abline(lm(yy[,1]~xx[,1]),col=4)

cor(x$mean_price,y$predict)^2

}
